refactor(archive): move phase semantics out of lib
