+++ devtools/build_topology_projection.py
@@ -106,7 +106,7 @@ LIB_PREFIX_TO_SUBPACKAGE = {
     "attribution": "lib/conversation/",
     "filter_": "archive/filter/",
     "filters": "archive/filter/",
-    "phase_": "lib/phase/",
+    "phase_": "archive/phase/",
     "projection_": "lib/projection/",
     "projections": "lib/projection/",
     "provider_": "lib/provider/",
@@ -211,7 +211,7 @@ TARGET_TO_OWNER = [
     ("polylogue/lib/message/", "lib-message"),
     ("polylogue/lib/conversation/", "lib-conversation"),
     ("polylogue/lib/semantic/", "lib-semantic"),
-    ("polylogue/lib/phase/", "lib-phase"),
+    ("polylogue/archive/phase/", "archive-phase"),
     ("polylogue/lib/projection/", "lib-projection"),
     ("polylogue/lib/provider/", "lib-provider"),
     ("polylogue/archive/attachment/", "archive-attachment"),
