+++ docs/topology-status.md
@@ -14,12 +14,13 @@ Generated by `devtools render-topology-status`. Reads `docs/plans/topology-targe
 | archive-action-event | — | 4 | 4 | 0 | 0 |
 | archive-attachment | — | 2 | 2 | 0 | 0 |
 | archive-filter | archive filter semantics | 4 | 4 | 0 | 0 |
+| archive-phase | — | 2 | 2 | 0 | 0 |
 | archive-query | archive query semantics | 17 | 17 | 0 | 0 |
 | archive-raw-payload | — | 5 | 5 | 0 | 0 |
 
 ### Summary
 
-- **Stable** (no move scoped): 626
+- **Stable** (no move scoped): 624
 - **Kernel** (polylogue/ root): 10
 - **Primitives** (lib-root + storage-root): 29
 - **TBD** (cell needs explicit assignment): 0
