+++ polylogue/lib/session/runtime.py
@@ -5,9 +5,9 @@ from __future__ import annotations
 from dataclasses import replace
 from typing import TYPE_CHECKING
 
+from polylogue.archive.phase.extraction import extract_phases
 from polylogue.lib.conversation.attribution import extract_attribution
 from polylogue.lib.conversation.extraction import extract_work_events
-from polylogue.lib.phase.extraction import extract_phases
 from polylogue.lib.semantic.facts import build_conversation_semantic_facts
 from polylogue.lib.session.models import SessionAnalysis, SessionProfile
